healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: d7f8e6ece8c1e47928acc7db51ecc5fae7e5446b
Former-commit-id: 589ede3ba1d09a5803160d8cfcdeaf76d84193d7
Former-commit-id: 144071b406db3a8227dbecc936ddb657a72d1f7c
Former-commit-id: f8438a520da49daed61595adda6cc55e8a25dcaa
Former-commit-id: 44d7caf9d0514aa75b209bc7653f8426e50a5bef