relax /system/bin directory permissions
In commit 664ccf7802de2d3862c5386f48c12f9c40e67836, we tightened the
permissions on various bin directories. Please see
https://android-review.googlesource.com/c/platform/system/core/+/822955
for details.

This change causes the Chase banking app to crash. This is because
the Chase app is using inotify_add_watch() on the /system/bin directory
and not checking the return value.

The Android Security model guarantees the immutability of files in
/system/bin, so the inotify watch is unnecessary.

Until the Chase app fixes their bug, we need to relax the permissions on
the /system/bin directory. Conceptually, this is a partial revert of
664ccf7802de2d3862c5386f48c12f9c40e67836.

Bug: 119605322
Test: compiles
Change-Id: Ic72dd24cb27cff677093963bdfd0ae09bf132e08
(cherry picked from commit acd5bf1d81b23fb205a3c564e0accabbd9fa110a)