Lengthen ScopedDisableMallocTimeout timeout to 10 seconds
memunreachable_unit_test is flaking sometimes with timeouts in
ScopedDisableMallocTimeout.  They can't be real deadlocks, which
is what this timeout is designed to detect, the Enable() called
from the signal handler wasn't doing anything because disabled_
hadn't been set yet.

Lengthen the timeout to 10 seconds, and set disabled_ before
starting to take the malloc locks in malloc_disable.  If this
really deadlocked then calling malloc_enable inside malloc_disable
would make a mess of the locks, but it would at least unlock
whatever lock was deadlocked and give the test a chance to
report the error.

Bug: 141229513
Test: atest memunreachable_unit_test
Change-Id: I3578964577025aaa4bbba09027afd22997d4adbd